add updateUser action
