+++ app/controllers/api/v1/auth/registrations_controller.rb
@@ -9,7 +9,7 @@
         end
 
         def account_update_params
-          params.permit(:name, :email)
+          params.permit(:name, :email, :age, :gender, :password, :password_confirmation)
         end
 
   #     end
